% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_NACE.R
\name{check_NACE}
\alias{check_NACE}
\title{Check NACE}
\usage{
check_NACE(df, ths_abs = 1, ths_rel = 0.5)
}
\arguments{
\item{df}{the data frame with the data. It should have the following column:TOTAL,A,BTE,F,GTI,GTJ,J,K,L,M_N,KTN,OTQ,RTU,OTU}

\item{ths_abs}{absolute threshold. Differences below the threshold are ignored.}

\item{ths_rel}{relative threshold. Differences below the threshold are ignored (1=1\%)}
}
\value{
a dataframe. If all pass are passed it will be empty.
}
\description{
Check NACE
}
\examples{
library(tidyverse)
df <- regacc_load_csv(
folder = "D:/data/REGACC/csv",
country_sel = c("AT"),
time_min = "2021-12-01"
) \%>\%
 dplyr::filter(table_identifier == "T1002") \%>\%
 dplyr::select(ref_area, sto, activity, unit_measure, time_period, obs_value) \%>\%
 tidyr::pivot_wider(
   names_from = activity,
   values_from = obs_value
 ) \%>\%
 rename(TOTAL = "_T")

check_NACE(df, ths_abs = 1, ths_rel = 0.5)
}
