% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_parquet.R
\name{csv_to_parquet}
\alias{csv_to_parquet}
\title{A function to convert csv files to parquet files}
\usage{
csv_to_parquet(dir)
}
\arguments{
\item{dir}{}
}
\value{
csv files are read with data.table and converted to parquet with arrow. The filenames are the same but replacing csv with parquet
}
\description{
A function to convert csv files to parquet files
}
\examples{

csv_to_parquet("where_my_csv_are_stored")
}
