% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_eurostat.R
\name{eurostat_cols}
\alias{eurostat_cols}
\title{Function to extract  colors as hex codes}
\usage{
eurostat_cols(...)
}
\arguments{
\item{...}{Character names of colors}
}
\description{
Function to extract  colors as hex codes
}
