% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_eurostat.R
\name{eurostat_pal}
\alias{eurostat_pal}
\title{Return function to interpolate a  color palette}
\usage{
eurostat_pal(palette = "new", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in luis_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a  color palette
}
