% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eurostat.R
\name{find_eurostat_dataset}
\alias{find_eurostat_dataset}
\title{Find Eurostat datasets codes}
\usage{
find_eurostat_dataset(x)
}
\arguments{
\item{x}{a string with *'s}
}
\value{
a data frame
}
\description{
Find Eurostat datasets codes
}
\examples{
df<-find_eurostat_dataset("nama_10r*")

}
