% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP.R
\name{get_annual_GDP}
\alias{get_annual_GDP}
\title{Get a dataframe with the latest Eurobase annual GDP}
\usage{
get_annual_GDP(na_item_sel = "B1GQ", unit_sel = "CP_MEUR")
}
\arguments{
\item{na_item_sel}{which na_item to download, B1GQ by default}

\item{unit_sel}{which unit, CP_MEUR by default}
}
\value{
a data.frame/data.table
}
\description{
This function downloads from Eurobase the latest GDP data and other variables in the table nama_10_gdp for the unit declared by the user.
}
\examples{
dt<- get_annual_GDP(na_item_sel=c("B1GQ", "D1"), unit_sel= c("CP_MEUR","CP_MNAC"))
}
