% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP.R
\name{get_quarterly_GDP}
\alias{get_quarterly_GDP}
\title{Get a dataframe with the latest Eurobase quarterly GDP}
\usage{
get_quarterly_GDP(
  na_item_sel = "B1GQ",
  unit_sel = "CLV_PCH_PRE",
  s_adj_sel = "SCA",
  min_time = "2019-01-01"
)
}
\arguments{
\item{na_item_sel}{which na_item to download, B1GQ by default}

\item{unit_sel}{which unit, CLV_PCH_PRE by default}

\item{s_adj_sel}{which adjustment, SCA by default}

\item{min_date}{from which year, 2020 by default}
}
\value{
a data.frame/data.table
}
\description{
This function downloads from Eurobase the latest quarterly GDP data and other variables in the table namq_10_gdp for the unit and adjustment declared by the user.
}
\examples{
dt<- get_quarterly_GDP(na_item_sel=c("B1GQ"), unit_sel= c("CP_MEUR","CP_MNAC"), s_adj_sel = "SCA")
}
