% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NQR_data.R
\docType{data}
\name{NQR_data}
\alias{NQR_data}
\title{NUTS 2021 classification}
\format{
A data.frame / data.table:
\describe{
\item{Country}{Country code}
\item{NUTS}{NUTS level:0 and 2}
\item{geo}{Code of the region (AT11,BE100,...)}
\item{na_item}{Code of the transaction (B1G,EMP,D1,B6N)}
\item{vintage}{year the data was received}
\item{time}{reference oeriod}
\item{valuesl}{values}
}
}
\source{
\url{internal}
}
\usage{
NQR_data

NQR_data(x, df1, df2, ncp, log = FALSE)
}
\description{
A data frame with the NUTS 2021 codes and labels.
}
\keyword{datasets}
