% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_linechart.R
\name{regacc_linechart}
\alias{regacc_linechart}
\title{Create standard line charts}
\usage{
regacc_linechart(
  dat,
  hor,
  ver,
  grp,
  clr,
  leg = TRUE,
  fac,
  fsz = 14,
  ffa = "mono",
  lsz = 0.8,
  hbr = 3,
  vbr = 4,
  acc = 1,
  exm = 0,
  exa = 0
)
}
\arguments{
\item{dat}{the data frame with the data}

\item{hor}{the name of the variable in the horizontal axis}

\item{ver}{the name of the variable in the vertical axis}

\item{grp}{the variable to use for grouping the data}

\item{clr}{the variable mapped to the colour aesthetic}

\item{leg}{set to FALSE if you do not want to show the legend}

\item{fac}{the (unique) variable to use if small multiples are wanted}

\item{fsz}{font size}

\item{ffa}{font family ("mono","sans", "serif) should be available in any EC laptop}

\item{lsz}{line size}

\item{hbr}{parameter for the number of horizontal breaks, input for scales::break_pretty}

\item{vbr}{parameter for the number of vertical breaks, input for scales::break_pretty}

\item{acc}{number of decimals for Y axis (0.1 means one decimal)}

\item{exm}{multiplicative parameter for the scale limits}

\item{exa}{additive parameter for the scale limits}
}
\value{
a ggplot
}
\description{
This functions makes easy to create standard line charts for different purposes
}
\examples{
df<- NQR_data \%>\%
filter(Country =="AT" & vintage =="2021" & NUTS =="2") \%>\%
  mutate(time=as.integer(time))

regacc_linechart(dat=df,
                 hor=time,
                 ver=values,
                 grp=geo,
                 clr=geo,
                 leg=TRUE,
                 fac=na_item)
}
