% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_load_all_csv.R
\name{regacc_load_all_csv}
\alias{regacc_load_all_csv}
\title{Load all xml cvs files}
\usage{
regacc_load_all_csv(folder, geo, time)
}
\arguments{
\item{folder}{path to the xml.csv files.}

\item{geo}{country(ies) to look for.}

\item{time}{date from which to take the files.}
}
\value{
a data frame
}
\description{
Loads the csv.xml files crated by Famefeed in a data frame.
}
\examples{
df<- regacc_load_all_csv(folder ="F:/data/REGACC/csv",
                         geo = c("MT", "LU"),
                         time = "2021-12-01")
}
