% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_load_MDT.R
\name{regacc_load_MDT}
\alias{regacc_load_MDT}
\title{Load files from MDT}
\usage{
regacc_load_MDT(folder, table)
}
\arguments{
\item{folder}{Path to the location of the files}

\item{table}{(Unique) table to load (nama_10r_2gdp, nama_10r_3gdp, nama_10r_3popgdp, nama_10r_3gva, nama_10r_3empers, nama_10r_2coe, nama_10r_2gfcf, nama_10r_2emhrw, nama_10r_2hhinc, nama_10r_2gvavr).}
}
\value{
a data frame/data.table object.
}
\description{
Loads the historical files that we keep from the internal copy of MDT.
The location of the files in the server is:
}
\examples{

df <- regacc_load_MDT(folder = "E:/data/REGACC/MDT/source",
                     table = "nama_10r_2gp")
}
