% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_load_sent_eurobase.R
\name{regacc_load_sent_eurobase}
\alias{regacc_load_sent_eurobase}
\title{Get files sent to Eurobase}
\usage{
regacc_load_sent_eurobase(
  folder,
  table,
  time_min,
  time_max = "2022-07-11",
  sel_country
)
}
\arguments{
\item{folder}{specifies the folder where the files are (zip files)}

\item{table}{the Eurobase table to look for. It only accepts a single argument (nama_10r_2gdp, nama_10r_3gdp, nama_10r_3popgdp, nama_10r_3gva, nama_10r_3empers, nama_10r_2coe, nama_10r_2gfcf, nama_10r_2emhrw, nama_10r_2hhinc, nama_10r_2gvavr)}

\item{time_min}{date of publication from where to start looking (yyyy-mm-dd)}

\item{time_max}{date of publication from where to stop looking (yyyy-mm-dd)}

\item{sel_country}{country or countries to look for}
}
\value{
a data frame
}
\description{
This function looks at the folder where the files sent to Eurobase are stored, loads them and puts them together in a data frame.
}
\examples{
df<- regacc_load_sent_eurobase(folder = "D:/data/REGACC/sent/source",
table="nama_10r_2gvagr",
time_min = "2020-01-01",
time_max = "2021-03-16",
sel_country =c("LU","EE"))
}
