% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_scatter.R
\name{regacc_scatter}
\alias{regacc_scatter}
\title{Create standard scatter charts}
\usage{
regacc_scatter(
  dat,
  hor,
  ver,
  grp,
  clr,
  leg = TRUE,
  fac,
  fsz = 14,
  ffa = "mono",
  psz = 0.8,
  hbr = 3,
  vbr = 4,
  acc1 = 1,
  acc2 = 1
)
}
\arguments{
\item{dat}{the data frame with the data}

\item{hor}{the name of the variable in the horizontal axis}

\item{ver}{the name of the variable in the vertical axis}

\item{clr}{the variable mapped to the colour aesthetic}

\item{leg}{set to FALSE if you do not want to show the legend}

\item{fac}{the (unique) variable to use if small multiples are wanted}

\item{fsz}{font size}

\item{ffa}{font family ("mono","sans", "serif) should be available in any EC laptop}

\item{psz}{point size}

\item{hbr}{parameter for the number of horizontal breaks, input for scales::break_pretty}

\item{vbr}{parameter for the number of vertical breaks, input for scales::break_pretty}

\item{acc1}{number of decimals for X axis (0.1 means one decimal)}

\item{acc2}{number of decimals for Y axis (0.1 means one decimal)}
}
\value{
a ggplot
}
\description{
This functions makes easy to create standard line charts for different purposes
}
\examples{
df<- NQR_data \%>\%
 filter(Country =="AT" & vintage =="2021" & na_item \%in\% c("B1G","B6N")) \%>\%
 pivot_wider(names_from = na_item,
             values_from = values)

regacc_scatter(dat=df,
              hor=B1G,
              ver=B6N,
              grp=geo,
              clr=geo,
              leg = FALSE,
              fac = geo,
              psz = 2)
}
