% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_vintages_gdp.R
\name{regacc_vintages_gdp}
\alias{regacc_vintages_gdp}
\title{Get regional GDP vintages from the dedicated section}
\usage{
regacc_vintages_gdp()
}
\value{
a data frame
}
\description{
The functions downloads the excel file in the dedicated section with the historical releases of regional GDP and makes it tidy.
}
\examples{
df<- regacc_vintages_gdp()
}
