% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_completeness.R
\name{report_completeness}
\alias{report_completeness}
\title{Report completeness as in NQR}
\usage{
report_completeness(data, table_sel)
}
\arguments{
\item{data}{a data frame with the reported data}

\item{table_sel}{individual table to check}
}
\value{
completeness ratios and missing series
}
\description{
The function provides the same ratios as reported in the NQR and an interactive table wit the missing values.
}
\examples{
df<- load_xml("D:/regacc/data/input/xml",
country_sel = "ES",
table_sel = c("T1001","T1002","T1200","T1300"),
time_min = "2021-10-01")

report_completeness (df)

}
