% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_eurostat.R
\name{scale_colour_eurostat}
\alias{scale_colour_eurostat}
\title{Color scale constructor for eurostat colors}
\usage{
scale_colour_eurostat(palette = "new", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in luis_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Color scale constructor for eurostat colors
}
