% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_eurostat.R, R/scale_theme2.R
\name{scale_fill_theme2}
\alias{scale_fill_theme2}
\title{Apply the new Eurostat palette to fill aesthetics in ggplot}
\usage{
scale_fill_theme2()

scale_fill_theme2()
}
\description{
Apply the new Eurostat palette to fill aesthetics in ggplot

Apply Eurostat Theme2 palette to fill aesthetics in ggplot
}
\examples{

ggplot2::ggplot(mtcars,aes(cyl,mpg, fill=as.factor(cyl)))+
geom_col()+
scale_fill_eurostat()


ggplot2::ggplot(mtcars,aes(cyl,mpg, fill=as.factor(cyl)))+
geom_col()+
scale_fill_theme2()

}
