% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_to_anything.R
\name{xlsx_to_anything}
\alias{xlsx_to_anything}
\title{A function to convert xlsx files to other formats}
\usage{
xlsx_to_anything(dir, extension, recursive = FALSE)
}
\arguments{
\item{dir}{}
}
\value{
xlsx files are converted with rio. The filenames are the same but replacing xlsx by the chosen format.
}
\description{
A function to convert xlsx files to other formats
}
\examples{

xlsx_to_anything("E:/test","csv",recursive = TRUE)
}
