% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_to_csv.R
\name{xlsx_to_csv}
\alias{xlsx_to_csv}
\title{A function to convert xlsx files to csv files}
\usage{
xlsx_to_csv(dir)
}
\arguments{
\item{dir}{}
}
\value{
xlsx files are converted with rio The filenames are the same but replacing xlsx by csv.
}
\description{
A function to convert xlsx files to csv files
}
\examples{

csv_to_parquet("where_my_csv_are_stored")
}
