% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eurostat.R
\name{find_eurostat_date}
\alias{find_eurostat_date}
\title{A function to get eurostat datasets updated after a certain date}
\usage{
find_eurostat_date(x)
}
\arguments{
\item{x}{a date in yyyy-mm-dd format}
}
\value{
a data frame
}
\description{
This function looks at eurostat data sets last updated date and filter those updated after the date provided by the user
}
\examples{
july_2022 <-find_eurostat_date("2022-07-01")
}
