% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eurostat.R
\name{find_eurostat_desc}
\alias{find_eurostat_desc}
\title{Find Eurostat datasets codes based on keywords in the description}
\usage{
find_eurostat_desc(x)
}
\arguments{
\item{x}{a string with *s}
}
\value{
a data frame
}
\description{
Find Eurostat datasets codes based on keywords in the description
}
\examples{
df <- find_eurostat_desc("*GDP*")
}
