% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eurostat.R
\name{get_eurostat_label_codes}
\alias{get_eurostat_label_codes}
\title{get most common national accounts dictionaries
Get in a dataframe the labels of the most common dictionaries used in national accounts}
\usage{
get_eurostat_label_codes()
}
\value{
a dataframe
}
\description{
This function collects several Eurostat dictionaries frequently used in National Accounts. The main purpose is to get the labels of the codes.
}
\examples{
df<- get_eurostat_label_codes()

}
