% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luispack-package.R
\docType{package}
\name{luispack-package}
\alias{luispack}
\alias{luispack-package}
\title{luispack: Set of personal functions that are handy}
\description{
Ggplot palettes, get eurostat data, export and convert utilities
}
\keyword{internal}
