% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS_2021.R
\docType{data}
\name{NUTS_2021}
\alias{NUTS_2021}
\title{NUTS 2021 classification}
\format{
A data.frame / data.table:
\describe{
\item{country}{Country code}
\item{NUTS}{NUTS level:0,1,2 and 3}
\item{geo}{Code of the region (AT11,BE100,...)}
\item{label}{Name of the region}
}
}
\source{
\url{https://ec.europa.eu/eurostat}
}
\usage{
NUTS_2021
}
\description{
A data frame with the NUTS 2021 codes and labels.
}
\keyword{datasets}
