% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_linechart.R
\name{regacc_linechart}
\alias{regacc_linechart}
\title{Create standard line charts}
\usage{
regacc_linechart(dat, hor, ver, grp, clr, leg = TRUE, fac)
}
\arguments{
\item{dat}{the data frame with the data}

\item{hor}{the name of the variable in the horizontal axis}

\item{ver}{the name of the variable in the vertical axis}

\item{grp}{the variable to use for grouping the data}

\item{clr}{the variable mapped to the colour aesthetic}

\item{leg}{set to FALSE if you do not want to show the legend}

\item{fac}{the (unique) variable to use if small multiples are wanted}
}
\value{
a ggplot
}
\description{
This functions makes easy to create standard line charts for different purposes
}
\examples{
df<- NQR_data \%>\%
filter(Country =="AT" & vintage =="2021" & NUTS =="2") \%>\%
  mutate(time=as.integer(time))

regacc_linechart(dat=df,
                 hor=time,
                 ver=values,
                 grp=geo,
                 clr=geo,
                 leg=TRUE,
                 fac=na_item)
}
