% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc_load_all_csv.R
\name{regacc_load_all_csv}
\alias{regacc_load_all_csv}
\title{Load all xml cvs files}
\usage{
regacc_load_all_csv(
  folder = "//fame2prod.cc.cec.eu.int/fame-estat/econ/REGACC/INSPACE",
  geo,
  tables = c("T1001", "T1002", "T1200", "T1300"),
  time_min = "2020-01-01",
  time_max = "2099-12-31"
)
}
\arguments{
\item{folder}{path to the xml.csv files.}

\item{geo}{country(ies) to look for.}

\item{tables}{tables to look for.}

\item{time_min}{file creation should be above this limit(yyyy-mm-dd).}

\item{time_max}{file creation should be below this limit(yyyy-mm-dd).}
}
\value{
a data frame
}
\description{
The objective of this function is to load several \emph{xml.csv} files stored in \href{"//fame2prod.cc.cec.eu.int/fame-estat/econ/REGACC/INSPACE"}{fame2prod.cc.cec.eu.int/fame-estat/econ/REGACC/INSPACE}. They are created after xml files are loaded with FameFeed.
The function needs these packages: \code{{data.table}}, \code{{tidyverse}} and \code{{janitor}}.
The function has several arguments:
\itemize{
\item folder: the path to the files. Normally it will be \href{"//fame2prod.cc.cec.eu.int/fame-estat/econ/REGACC/INSPACE"}{fame2prod.cc.cec.eu.int/fame-estat/econ/REGACC/INSPACE} and is set as the default option that is used if no argument is provided.
\item geo: the countries to look at. It accepts one ("LU") or several c("LU","MT") arguments. No argument is provided by default.
\item tables: the table ("T1001") or tables c("T1200","T1300") to look for. All tables are selected by default,
\item time_min: the function looks at the time stamp of the files. We can filter for files created after a certain date. The default option is ""2020-01-01".
\item time_max: We can filter for files created before a certain date. The default option is ""2099-12-31".
}
}
\examples{
regacc_load_all_csv(folder ="E:/data/REGACC/csv",
                   geo = c("MT", "LU"),
                   time_min = "2021-12-01")
}
