% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_eurostat.R
\name{scale_colour_eurostat}
\alias{scale_colour_eurostat}
\title{Apply the new Eurostat palette to colour aesthetics in ggplot}
\usage{
scale_colour_eurostat()
}
\description{
Apply the new Eurostat palette to colour aesthetics in ggplot
}
