% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_theme2.R
\name{scale_colour_theme2}
\alias{scale_colour_theme2}
\title{Apply Eurostat Theme2 palette to colour aesthetics in ggplot}
\usage{
scale_colour_theme2()
}
\description{
Apply Eurostat Theme2 palette to colour aesthetics in ggplot
}
\examples{
ggpplot2::ggplot(mtcars,aes(cyl,mpg, colour=as.factor(cyl)))+
geom_point(size=3)+
  scale_colour_theme2()
}
