% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_eurostat.R
\name{scale_fill_eurostat}
\alias{scale_fill_eurostat}
\title{Apply the new Eurostat palette to fill aesthetics in ggplot}
\usage{
scale_fill_eurostat()
}
\description{
Apply the new Eurostat palette to fill aesthetics in ggplot
}
\examples{
library(tidyverse)
ggplot(mtcars,aes(cyl,mpg, fill=as.factor(cyl)))+
geom_col()+
scale_fill_eurostat()

}
