% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_to_anything.R
\name{xlsx_to_anything}
\alias{xlsx_to_anything}
\title{Convert xlsx files to other formats}
\usage{
xlsx_to_anything(dir, extension, recursive = FALSE)
}
\arguments{
\item{dir}{older in which the excel files to be converted are stored}

\item{extension}{extension to which the files will be converted (for example,".csv")}

\item{recursive}{whether lo kook recursively in sub-folders, FALSE by default}
}
\value{
xlsx files are converted with rio. The filenames are the same but replacing xlsx by the chosen format.
}
\description{
Convert xlsx files to almost any other format. Works in sub-folders if recursive is set to TRUE. Original files are not deleted in case they want to be kept and for precautionary reasons.
}
\examples{

xlsx_to_anything("E:/test",".csv",recursive = TRUE)
}
